src/data_structure/easy/_002_2darray/Result.java
package data_structure.easy._002_2darray;

import java.util.List;

public class Result {

    /*
     * Complete the 'hourglassSum' function below.
     *
     * The function is expected to return an INTEGER.
     * The function accepts 2D_INTEGER_ARRAY arr as parameter.
     */

    public static int hourglassSum(List<List<Integer>> arr) {
        boolean isfirst = true;
        int max = 0;
        for (int i = 0; i < 4; i++) {
            for (int j = 0; j < 4; j++) {
                int sum = arr.get(i).get(j) + arr.get(i).get(j + 1) + arr.get(i).get(j + 2) +
                        arr.get(i + 1).get(j + 1) +
                        arr.get(i + 2).get(j) + arr.get(i + 2).get(j + 1) + arr.get(i + 2).get(j + 2);
                if(!isfirst) {
                  if(max < sum) {
                      max = sum;
                  }
                } else {
                    max = sum;
                    isfirst = false;
                }
            }
        }
        return max;
    }

}
